+++ tests/test_skool2html.py
@@ -139,6 +139,9 @@ def mock_run(*args):
     global run_args
     run_args = args
 
+def mock_write_disassembly(*args):
+    pass
+
 class TestHtmlWriter(HtmlWriter):
     def init(self):
         global html_writer
@@ -190,6 +193,14 @@ class TestHtmlWriter(HtmlWriter):
 
     def write_index(self, *args):
         self.add_call('write_index', args)
+
+class MockImageWriter:
+    def __init__(self, palette, options):
+        global mock_image_writer
+        self.palette = palette
+        self.options = options
+        self.default_format = None
+        mock_image_writer = self
 
 class Skool2HtmlTest(SkoolKitTestCase):
     def setUp(self):
@@ -352,6 +363,39 @@ class Skool2HtmlTest(SkoolKitTestCase):
             self.assertEqual(error, '')
             self.assertEqual(output[0], 'Using skool file: {0}'.format(skoolfile))
             self.assertTrue(isfile(join(self.odir, game_dir, 'asm', '40000.html')))
+
+    def test_colour_parsing(self):
+        self.mock(skool2html, 'write_disassembly', mock_write_disassembly)
+        self.mock(skool2html, 'ImageWriter', MockImageWriter)
+
+        # Valid colours
+        exp_colours = (
+            ('RED', '#C40000', (196, 0, 0)),
+            ('WHITE', '#cde', (204, 221, 238)),
+            ('YELLOW', '198,197,0', (198, 197, 0))
+        )
+        colours = ['[Colours]']
+        colours.extend(['{}={}'.format(name, spec) for name, spec, rgb in exp_colours])
+        reffile = self.write_text_file('\n'.join(colours), suffix='.ref')
+        self.write_text_file(path='{0}.skool'.format(reffile[:-4]))
+        output, error = self.run_skool2html(reffile)
+        self.assertEqual(error, '')
+        for name, spec, rgb in exp_colours:
+            self.assertEqual(mock_image_writer.palette[name], rgb)
+
+        # Invalid colours
+        bad_colours = (
+            ('BLACK', ''),
+            ('CYAN', '#)0C6C5'),
+            ('MAGENTA', '!98,0,198')
+        )
+        for name, spec in bad_colours:
+            colours = ['[Colours]', '{}={}'.format(name, spec)]
+            reffile = self.write_text_file('\n'.join(colours), suffix='.ref')
+            self.write_text_file(path='{0}.skool'.format(reffile[:-4]))
+            with self.assertRaises(SkoolKitError) as cm:
+                self.run_skool2html(reffile)
+            self.assertEqual(cm.exception.args[0], 'Invalid colour spec: {}={}'.format(name, spec))
 
     def test_option_w(self):
         options = [